Move label type to shared types
