+++ application/ui/src/constants/shared-types.ts
@@ -1,0 +1,6 @@
+// Copyright (C) 2025 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+import { components } from 'src/api/openapi-spec';
+
+export type Label = components['schemas']['LabelView'];
