+++ application/ui/src/features/annotator/annotations/annotation-labels.component.test.tsx
@@ -2,10 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { getMockedLabel } from 'mocks/mock-labels';
+import { Label } from 'src/constants/shared-types';
 import { fireEvent, render, screen } from 'test-utils/render';
 import { describe, expect, it, vi } from 'vitest';
 
-import { Label } from '../types';
 import { AnnotationLabels } from './annotation-labels.component';
 
 const mockZoom = { scale: 1, maxZoomIn: 10, hasAnimation: false, translate: { x: 0, y: 0 } };
