+++ application/ui/src/features/annotator/annotations/annotation-shape-with-labels.component.tsx
@@ -4,10 +4,11 @@
 import { Key } from 'react';
 
 import polylabel from 'polylabel';
+import { Label } from 'src/constants/shared-types';
 
 import { useAnnotationActions } from '../annotation-actions-provider.component';
 import { useAnnotationVisibility } from '../annotation-visibility-provider.component';
-import { Annotation, Label, Polygon } from '../types';
+import { Annotation, Polygon } from '../types';
 import { AnnotationLabels } from './annotation-labels.component';
 import { AnnotationShape } from './annotation-shape.component';
