+++ application/ui/src/features/project/label-selection/interface.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Label } from '../../annotator/types';
+import { Label } from 'src/constants/shared-types';
 
 export type LabelItemProps = {
     label: Label;
