+++ application/ui/src/features/project/list/project-card.component.tsx
@@ -6,6 +6,7 @@ import { clsx } from 'clsx';
 import { NavLink } from 'react-router-dom';
 
 import { SchemaProjectView } from '../../../api/openapi-spec';
+// TODO: replace mock thumbnail once /api/projects/{project_id}/thumbnail is finished
 import thumbnailUrl from '../../../assets/mocked-project-thumbnail.png';
 import { paths } from '../../../constants/paths';
 import { MenuActions } from './menu-actions.component';
